Modified The EMACalculator
Added in a flag for 70 and 140.
Refined Logic in the EMAcalculator.
Updated the CSV with newer values now. It's also in the right order, added more accurate 2,21,70 calculations which were previously only averages of each until like 140

Next Commit will be to our Dollal Function, I'll try and push it next today.